requirements.txt
Django==2.0.4
django-jet==1.0.7
djangorestframework==3.8.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.8.0
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
